TriggerMatchingTool: Fix compilation warnings.
Remove unused variable.
Fix signed/unsigned comparison.



Former-commit-id: f47febc0c8bd7c0d17ae55ff561959f4435010a6

+++ Trigger/TrigAnalysis/TriggerMatchingTool/Root/MatchingTool.cxx
@@ -140,7 +140,7 @@ namespace Trig {
     
 
     //record sum of costs
-    for(int i = 0; i < result.assignment.size(); ++i){
+    for(size_t i = 0; i < result.assignment.size(); ++i){
       result.total_cost += distance_matrix[i][result.assignment[i]];
     }
